Refer to SVG files using Javascript URL
At some point in the past I added the SVG files to
the 'parcel serve' command in tasks.json so they
were added to the .parcel-cache folder. Later I
removed them from the command because it appeared
to not be needed. But when I deleted the
.parcel-cache folder, the SVGs would not work
because the HTML file does not refer to them so
Parcel doesn't know to bundle them.