Move common CSS for overlay popup to separate file
In the operator window, there are currently three
popups: special category, buzz history chart, and
game-end charts.

In the presentation window the only popup is the
special category popup, while buzz history chart and
game-end charts are shown as presentation slides.

The presentation window has transitions for the
special category popup and backdrop, but the operator
window does not have any transitions.

Previously, in the operator window, the backdrop and
all the popups were shown/hidden by setting display
to "block" or "none". To use the same CSS in both
windows, I changed it so when operator popups are not
shown they are translated off-screen.

I think this might not be great. So I think I should
either not use any popups in either window, or move
stuff from presentation slides to popups.