Allow showing question on screen with the answer
Previously, to respond to keyboard input in the state
machine without changing states, you added a keyboard
transition with the destination state set to the origin
state. If the state had an ON_ENTER function it would be
called every time the keyboard transition happened.

The "showAnswer" state has an ON_ENTER function which
does a bunch of stuff, so I did not want to repeatedly
call that. So I added keyboard transition to the state
machine, which respond to keyboard events without
changing states.